Update for new team builder draft mode
Riot broke the script with the new team builder mode because the queue type was different. Initially, it was just completely broken because it failed to update on any games. It's fixed now, the queue type is just different now. Note that the queue type of TEAM_BUILDER_DRAFT_RANKED_5x5 returns games under the old system marked as solo queue now, whereas the old value does not return new team builder games.