Fixes for issues #5 and #6
For issue #5, the code has been removed entirely and it will no longer check. Please refer to the issue for more details. This is the ability to check duoers, so that has been removed entirely.

Issue #6 is the ability to remove certain columns. It has been added, but some columns are mandatory and some columns rely on other columns. Please refer to the issue to see more details about these cells.