Update error codes to be more informative
Recently Riot has been having problems with their match history. Legitimate games simply aren't showing up on the match history until long after they are completed. For some strange reason, the matchlist endpoint correctly returns the match ID, but the match endpoint returns 404 on the match ID. The script never considered this possibility since it was never a reality, and so it always just returned "please make sure everything is configured correctly" as if to suggest it was a user problem rather than a Riot problem.

The script now checks for error code 404, which is not found. This happens in two scenarios, where the summoner ID couldn't be found or the game couldn't be found. If it's a case of summoner ID, it'll prompt a message for the user to check it's correct. If it's not the summoner ID (validated by whether the summoner ID is present or not, as it must be in order to make API calls), then it's a Riot problem and will return an appropriate message.

Note that at this time, I am unsure how long this issue will be going on for. Riot's pointed out they've been having problems, but they're happening again. Like bans, it is a Riot issue that they are aware of, working on, and out of my control. The first time the issue happened, I don't think it took more than a day to resolve. Like the error message suggests, I'll look into lasting problems to see if there's maybe a different issue. If there isn't, I'll look into a possible workaround (not that I think there is such a thing)